split more methods
